+++ src/main/java/flagship/domain/utils/tariffs/LightDueTariff.java
@@ -7,6 +7,7 @@ import lombok.Setter;
 import org.javatuples.Pair;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Getter
@@ -15,10 +16,10 @@ import java.util.Map;
 @NoArgsConstructor
 public class LightDueTariff extends Tariff {
 
-  private Map<Pair<Integer, Integer>, Double> lightDuesByGrossTonnage;
-  private Map<ShipType, Double> lightDuesPerTonByShipType;
-  private Map<ShipType, Double> discountCoefficientsByShipType;
-  private Integer lightDueMaximumValue;
+  private Map<Pair<Integer, Integer>, BigDecimal> lightDuesByGrossTonnage;
+  private Map<ShipType, BigDecimal> lightDuesPerTonByShipType;
+  private Map<ShipType, BigDecimal> discountCoefficientsByShipType;
+  private BigDecimal lightDueMaximumValue;
   private Integer callCountThreshold;
-  private Double callCountDiscountCoefficient;
+  private BigDecimal callCountDiscountCoefficient;
 }
